leet

+++ ARRAYS/MaxSubArray/brute.java
@@ -1,12 +1,12 @@
 public class brute {
     public static void main(String[] args) {
         int[] arr = {2, 3, 5, 1, 9};
-        long k = 10;
+        int k = 10;
         int len = getLongestSubarray(arr, k);
         System.out.println("The length of the longest subarray is: " + len);
     }
 
-    static int getLongestSubarray(int[] arr, long k) {
+    static int getLongestSubarray(int[] arr, int k) {
         int n = arr.length;
         int maxlen = 0;
